kv260 config: WIP config fragments for Linux v6.6
These config fragments are for use with the v6.6[.18] kernel.
They go in arch/arm64/configs and are used on top of defconfig.

$ make defconfig zynqmp-extras.config

Despite being here they are not yet suitable for use with linux-yocto.
Some settings would cause errors or warnings and extra configs will be
needed that we are currently getting from defconfig.

zynqmp-extras.config is my corrent best guest of the extra configs needed.
Some stuff for other Xilinx platfroms may be included and some =y might be
able to be =m's.

HDMI output does not work with the zynqmp-extras.config.
The module tries to probe but gets the following error:
[    1.850166] zynqmp-dpsub fd4a0000.display: DP output port not connected
[    1.856814] zynqmp-dpsub: probe of fd4a0000.display failed with error -22

zynqmp-display.config is my shotgub attempt to get HDMI working based on
the config settings present in Xilinx v2023.2 (6.1.30-xilinx-v2023.2)

It can be used alone or with zynqmp-extras but it makes no difference.
The error is the same.

Signed-off-by: Bill Mills <bill.mills@linaro.org>