kconfig: checkin current baselines and fragments
Check-in the current reference baselines and the zynqmp-* config fragments.

ross-v1 and ross-v2 are actually a resulting .config from building with
specific version of yocto-kernel-cache from
https://gitlab.com/rossburton/yocto-kernel-cache

ross-v1 is commit 0c44cfe68aa8169c1e6c860c2ab1ae0e865e44f2
* This is before the redo to better match defconfig.

ross-v2 is commit 03628b9a0e6739ee8a210c8e90921ac68823eb3f
* This is after the redo to be "inspired by" defconfig.

zynqmp-ross should be completely redundant when added to ross-v2.

zynqmp-defconfig should be completely redundant when added to defconfig
but not when addded to ross-v2.

zynqmp-new is the new stuff we need to add that is not in ross-v2 nor
defconfig.

make allnoconfig ross-v2.config zynqmp-defconfig.config zynqmp-new.config

Builds a kernel that works the same as the kv260 baseline of March 6.

The others zynqmp-{display,sound,v4l2} are WIP and overkill to see if we
can get something that works for these areas.  Nothing works yet.

Signed-off-by: Bill Mills <bill.mills@linaro.org>