kv260: HDMI and DP works
To fix HDMI (and DisplayPort) on kv260 w/ 6.6.20, two things are needed:
* An updated DTB that declares and wires up DP connectors
* kconfig that includes CONFIG_DRM_DISPLAY_CONNECTOR

The included patch for v6.6.20 Xilinx dts does the first thing.
The update to zynqmp-defconfig does the 2nd.

CONFIG_DRM_DISPLAY_CONNECTOR is not mentioned explicitly in v6.6 defconfig
but it is selected by the MESON DRM module so it end up in the config.
If we don't include MESON DRM, we need to explicitly ask for the connector.

Signed-off-by: Bill Mills <bill.mills@linaro.org>